Improve CMakeLists definition of font directory
The definition of the font directory was using CMAKE_SOURCE_DIR which is
not robust when the project is restructured or incorporated into other
project hierarchies. Use CMAKE_CURRENT_LIST_DIR instead.

The approval tests project was still setting the font directory even
though all the font handling was moved to the fonts_for_tests directory.
Remove the setting of the preprocessor directive from the approval tests
project's CMakeLists file.