Add font size argument to layout API
Until this commit the font sizes were hard-wired with display and text
mode fonts at 10 points, script at 7 points and script script at 5
points. Now the caller must provide the size of the display/text mode
fonts and script and script script will be 70% and 50% of that provided
size respectively.