Add total width and height to result
It is very useful for callers of the layout function to know the total
width and height of the layout. Although this information is implicitly
contained in the glyphs and lines that already make up the result, it is
not trivial to compute. Seeing as the layout process already generates
the total width and height as a side effect, it makes sense to return it
along with the rest of the result.